PWA-915 Migrate new address management to new router (@shopgate/user)
