+++ themes/theme-gmd/components/NavDrawer/components/StoreInfo/components/ImprintButton/index.jsx
@@ -1,40 +1,40 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import {
-  NAV_MENU_TERMS,
-  NAV_MENU_TERMS_AFTER,
-  NAV_MENU_TERMS_BEFORE,
+  NAV_MENU_IMPRINT,
+  NAV_MENU_IMPRINT_AFTER,
+  NAV_MENU_IMPRINT_BEFORE,
 } from '@shopgate/pwa-common/constants/Portals';
 import { PAGE_PATH } from '@shopgate/pwa-common/constants/RoutePaths';
 import Portal from '@shopgate/pwa-common/components/Portal';
-import DescriptionIcon from '@shopgate/pwa-ui-shared/icons/DescriptionIcon';
+import InfoIcon from '@shopgate/pwa-ui-shared/icons/InfoIcon';
 import { NavDrawer } from '@shopgate/pwa-ui-material';
-import connect from '../../connector';
+import connect from '../../../../connector';
 
-const LABEL = 'navigation.terms';
-const PAYMENT_PATH = `${PAGE_PATH}/terms`;
+const LABEL = 'navigation.about';
+const PAYMENT_PATH = `${PAGE_PATH}/imprint`;
 
 /**
  * @param {Function} props.navigate The navigate action.
  * @returns {JSX}
  */
-const TermsButton = ({ navigate }) => (
+const ImprintButton = ({ navigate }) => (
   <Fragment>
-    <Portal name={NAV_MENU_TERMS_BEFORE} />
-    <Portal name={NAV_MENU_TERMS}>
+    <Portal name={NAV_MENU_IMPRINT_BEFORE} />
+    <Portal name={NAV_MENU_IMPRINT}>
       <NavDrawer.Item
         label={LABEL}
-        icon={DescriptionIcon}
+        icon={InfoIcon}
         onClick={navigate(PAYMENT_PATH, LABEL)}
-        testId="navDrawerTermsButton"
+        testId="navDrawerImprintButton"
       />
     </Portal>
-    <Portal name={NAV_MENU_TERMS_AFTER} />
+    <Portal name={NAV_MENU_IMPRINT_AFTER} />
   </Fragment>
 );
 
-TermsButton.propTypes = {
+ImprintButton.propTypes = {
   navigate: PropTypes.func.isRequired,
 };
 
-export default connect(TermsButton);
+export default connect(ImprintButton);
