+++ themes/theme-gmd/components/NavDrawer/components/StoreInfo/components/ReturnsButton/index.jsx
@@ -9,7 +9,7 @@ import { PAGE_PATH } from '@shopgate/pwa-common/constants/RoutePaths';
 import Portal from '@shopgate/pwa-common/components/Portal';
 import DescriptionIcon from '@shopgate/pwa-ui-shared/icons/DescriptionIcon';
 import { NavDrawer } from '@shopgate/pwa-ui-material';
-import connect from '../../connector';
+import connect from '../../../../connector';
 
 const LABEL = 'navigation.return_policy';
 const PAYMENT_PATH = `${PAGE_PATH}/return_policy`;
