+++ themes/theme-gmd/components/NavDrawer/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import appConfig from '@shopgate/pwa-common/helpers/config';
-import showReturnPolicy from '@shopgate/pwa-common-commerce/market/helpers/showReturnPolicy';
 import ClientInformation from '@shopgate/pwa-ui-shared/ClientInformation';
 import { NavDrawer } from '@shopgate/pwa-ui-material';
 import Header from './components/Header';
@@ -9,12 +8,7 @@ import CategoryButton from './components/CategoryButton';
 import FavoritesButton from './components/FavoritesButton';
 import CartButton from './components/CartButton';
 import QuickLinks from './components/QuickLinks';
-import ShippingButton from './components/ShippingButton';
-import PaymentButton from './components/PaymentButton';
-import TermsButton from './components/TermsButton';
-import PrivacyButton from './components/PrivacyButton';
-import ReturnsButton from './components/ReturnsButton';
-import ImprintButton from './components/ImprintButton';
+import StoreInfo from './components/StoreInfo';
 import LogoutButton from './components/LogoutButton';
 
 // TODO: Add submenu header support
@@ -31,13 +25,7 @@ const NavDrawerContainer = () => (
     {appConfig.hasFavorites && <FavoritesButton />}
     <CartButton />
     <QuickLinks />
-    <NavDrawer.Divider />
-    <ShippingButton />
-    <PaymentButton />
-    <TermsButton />
-    <PrivacyButton />
-    {showReturnPolicy && <ReturnsButton />}
-    <ImprintButton />
+    <StoreInfo />
     <LogoutButton />
     <ClientInformation />
   </NavDrawer>
