+++ themes/theme-gmd/pages/index.jsx
@@ -32,6 +32,7 @@ import { APP_GLOBALS } from '@shopgate/pwa-common/constants/Portals';
 import SnackBar from 'Components/SnackBar';
 import Viewport from 'Components/Viewport';
 import View from 'Components/View';
+import AppBar from 'Components/AppBar/presets/BackBar';
 import Dialog from '@shopgate/pwa-ui-shared/Dialog';
 import locale from '../locale';
 import * as routes from './routes';
@@ -45,7 +46,7 @@ const devFontsUrl = 'https://fonts.googleapis.com/css?family=Roboto:400,400i,500
 const Pages = ({ store }) => (
   <App locale={locale} store={store}>
     <AppContext.Provider value={{ ...appConfig }}>
-      <ThemeContext.Provider value={{ View }}>
+      <ThemeContext.Provider value={{ View, AppBar }}>
         <LoadingProvider>
           <ToastProvider>
             <Portal name={APP_GLOBALS} />
